Move createRule to dedicated folder and sort tags bug
